+++ example/lib/define.dart
@@ -1,0 +1,5 @@
+class Define {
+  static const HAND_DETECTOR_MODEL = '/home/hedgehao/test/hand_landmark_lite.tflite';
+  static const FACE_DETECTOR_MODEL = '/home/hedgehao/test/faceDetector.tflite';
+  static const EFFICIENT_NET_MODEL = '/home/hedgehao/test/efficient_net.tflite';
+}