+++ example/lib/openni2/openni2.controller.dart
@@ -42,27 +42,15 @@ class OpenNIController extends GetxController {
   }
 
   void closeOpenNICamera() async {
-    if (cam == null) {
-      return;
-    }
-
-    await cam!.close();
+    await cam?.close();
   }
 
   void enableStreaming() async {
-    if (cam == null) {
-      return;
-    }
-
-    await cam!.enableStream();
+    await cam?.enableStream();
   }
 
   void disableStreaming() async {
-    if (cam == null) {
-      return;
-    }
-
-    await cam!.disableStream();
+    await cam?.disableStream();
   }
 
   void pipelineRGB() async {
@@ -77,12 +65,20 @@ class OpenNIController extends GetxController {
   void pipelineDepth() async {
     FvPipeline depthPipeline = cam!.depthPipeline;
     await depthPipeline.clear();
-    await depthPipeline.convertTo(0, 255.0 / 1024.0);
+    await depthPipeline.convertTo(OpenCV.CV_8U, 255.0 / 1024.0);
     await depthPipeline.applyColorMap(Random().nextInt(20));
     await depthPipeline.cvtColor(OpenCV.COLOR_RGB2RGBA);
     await depthPipeline.show();
 
     update([BUILDER_TEXTURE]);
+  }
+
+  void pipelineIr() async {
+    FvPipeline irPipeline = cam!.irPipeline;
+    await irPipeline.clear();
+    await irPipeline.convertTo(OpenCV.CV_8U, 255.0 / 1024.0);
+    await irPipeline.cvtColor(OpenCV.COLOR_RGB2RGBA);
+    await irPipeline.show();
   }
 
   void selectDevice(int selectIndex) {
@@ -121,5 +117,12 @@ class OpenNIController extends GetxController {
     openglTextureId = await FlutterVision.getOpenglTextureId();
 
     update([BUILDER_TEXTURE_OPENGL]);
+  }
+
+  Future<void> deconstruct() async {
+    await cam?.disableStream();
+    await cam?.close();
+    cam = null;
+    update([BUILDER_TEXTURE]);
   }
 }