+++ example/lib/realsense/realsense.controller.dart
@@ -55,27 +55,15 @@ class RealsenseController extends GetxController {
   }
 
   void closeRealsenseCamera() async {
-    if (cam == null) {
-      return;
-    }
-
-    await cam!.close();
+    await cam?.close();
   }
 
   void enableStreaming() async {
-    if (cam == null) {
-      return;
-    }
-
-    await cam!.enableStream();
+    await cam?.enableStream();
   }
 
   void disableStreaming() async {
-    if (cam == null) {
-      return;
-    }
-
-    await cam!.disableStream();
+    await cam?.disableStream();
   }
 
   void pipelineRGB() async {
@@ -90,10 +78,19 @@ class RealsenseController extends GetxController {
   void pipelineDepth() async {
     FvPipeline depthPipeline = cam!.depthPipeline;
     await depthPipeline.clear();
-    await depthPipeline.convertTo(0, 255.0 / 1024.0);
+    await depthPipeline.convertTo(OpenCV.CV_8U, 255.0 / 1024.0);
     await depthPipeline.applyColorMap(Random().nextInt(20));
     await depthPipeline.cvtColor(OpenCV.COLOR_RGB2RGBA);
     await depthPipeline.show();
+
+    update([BUILDER_TEXTURE]);
+  }
+
+  void pipelineIr() async {
+    FvPipeline irPipeline = cam!.irPipeline;
+    await irPipeline.clear();
+    await irPipeline.cvtColor(OpenCV.COLOR_RGB2RGBA);
+    await irPipeline.show();
 
     update([BUILDER_TEXTURE]);
   }
@@ -134,5 +131,12 @@ class RealsenseController extends GetxController {
     openglTextureId = await FlutterVision.getOpenglTextureId();
 
     update([BUILDER_TEXTURE_OPENGL]);
+  }
+
+  Future<void> deconstruct() async {
+    await cam?.disableStream();
+    await cam?.close();
+    cam = null;
+    update([BUILDER_TEXTURE]);
   }
 }